:green_heart: #29 Move to the "emergency.spec"
